add auth to todo routes

+++ routers/todos.py
@@ -5,6 +5,7 @@ from starlette import status
 import models
 from database import SessionLocal
 from pydantic import BaseModel,Field
+from .auth import get_current_user
 
 
 router = APIRouter()
@@ -19,6 +20,7 @@ def get_db():
         db.close()
 
 db_dependency= Annotated[Session , Depends(get_db)]
+user_dependency= Annotated[dict , Depends(get_current_user)]
 
 class TodoRequest (BaseModel):
     title:str = Field(min_length=3)
@@ -27,27 +29,38 @@ class TodoRequest (BaseModel):
     complete: bool
 
 @router.get("/",status_code=status.HTTP_200_OK)
-async def read_all(db :db_dependency ):
-    if db.query(models.Todos).all() is not None:
+async def read_all(user:user_dependency,db :db_dependency ):
+    if user is None :
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='Authentication failed')
+    #gets all todos for the authorized user
+    if db.query(models.Todos).filter(user.get('id')== models.Todos.owner_id).all() is not None:
         return db.query(models.Todos).all()
     raise HTTPException(status_code=404,detail="Todo not found")
 
 @router.get("/todo/{todo_id}",status_code=status.HTTP_200_OK)
-async def read_todo(db:db_dependency,todo_id:int=Path(gt=0)):
-  todo_model=  db.query(models.Todos).filter(models.Todos.id ==todo_id).first()
-  if todo_model is not None:
-      return todo_model
-  raise HTTPException(status_code=404,detail="بچو کره دو")
+async def read_todo(user:user_dependency,db:db_dependency,todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Authentication failed')
+    todo_model=db.query(models.Todos).filter(models.Todos.id ==todo_id)\
+        .filter(user.get('id')== models.Todos.owner_id).first()
+    if todo_model is not None:
+        return todo_model
+    raise HTTPException(status_code=404,detail="بچو کره دو")
 
 @router.post("/todos",status_code=status.HTTP_201_CREATED)
-async def create_todo(db:db_dependency,todo_request:TodoRequest):
-    todo_model=models.Todos(**todo_request.model_dump())
+async def create_todo(user:user_dependency,db:db_dependency,todo_request:TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='Authentication failed')
+    todo_model=models.Todos(**todo_request.model_dump(),owner_id=user.get('id'))
     db.add(todo_model)
     db.commit()
-    return "model created"
+
 @router.put("/todos/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async  def update_todo(db:db_dependency,todo_request:TodoRequest ,todo_id:int=Path(gt=0)):
-    todo_model=db.query(models.Todos).filter(models.Todos.id ==todo_id).first()
+async  def update_todo(user:user_dependency,db:db_dependency,todo_request:TodoRequest ,todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='Authentication failed')
+    todo_model=db.query(models.Todos).filter(models.Todos.id ==todo_id)\
+        .filter(models.Todos.owner_id==user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404,detail="Todo not found")
     todo_model.title=todo_request.title
@@ -59,9 +72,12 @@ async  def update_todo(db:db_dependency,todo_request:TodoRequest ,todo_id:int=Pa
     return "todo updated"
 
 @router.delete("/todos/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency, todo_id:int=Path(gt=0)):
-    todo_model=db.query(models.Todos).filter(models.Todos.id ==todo_id).first()
+async def delete_todo(user:user_dependency,db:db_dependency, todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Authentication failed')
+    todo_model=db.query(models.Todos).filter(models.Todos.id ==todo_id)\
+        .filter(models.Todos.owner_id==user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404,detail="Todo not found")
-    db.query(models.Todos).filter(models.Todos.id==todo_id).delete()
+    db.query(models.Todos).filter(models.Todos.id==todo_id).filter(models.Todos.owner_id==user.get("id")).delete()
     db.commit()
